+++ spec/integration/mysql2_master_slave_adapter_spec.rb
@@ -4,8 +4,8 @@ require 'rspec'
 require 'master_slave_adapter'
 require 'integration/helpers/shared_mysql_examples'
 
-describe "ActiveRecord::ConnectionAdapters::MysqlMasterSlaveAdapter" do
-  let(:connection_adapter) { 'mysql' }
+describe "ActiveRecord::ConnectionAdapters::Mysql2MasterSlaveAdapter" do
+  let(:connection_adapter) { 'mysql2' }
 
   it_should_behave_like "a MySQL MasterSlaveAdapter"
 end